requirements.txt
torch
torchvision
streamlit
pillow
scikit-learn
matplotlib
numpy
